main: move main logic for running lndk into a library
Move the logic for running lndk into a lib.rs so we can import it
in the integration tests.
